+++ compare-frontend/src/app/views/mappage/modal/modal.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core';
+import { Partition, ScenarioName } from 'src/app/state/state';
+
+@Component({
+  selector: 'app-modal',
+  templateUrl: './modal.component.html',
+  styleUrls: ['./modal.component.css']
+})
+export class ModalComponent {
+  @Input() scenario!: ScenarioName;
+  @Input() partition!: Partition;
+  @Input() data!: object;
+
+  constructor() {
+    console.log("Creating modal for ", this.partition);
+  }
+}