+++ compare-frontend/src/app/views/startpage/startpage.component.ts
@@ -2,9 +2,9 @@ import { ParseSourceFile } from '@angular/compiler';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { RiesgosState, Rules, ScenarioName } from 'src/app/state/state';
+import { RiesgosState, ScenarioName } from 'src/app/state/state';
 import * as AppActions from '../../state/actions';
-import { RuleService, RuleSetName } from 'src/app/services/rule.service';
+import { RuleSetName } from 'src/app/state/rules';
 
 @Component({
   selector: 'app-startpage',
@@ -19,7 +19,6 @@ export class StartpageComponent implements OnInit {
 
   constructor(
     private store: Store<{riesgos: RiesgosState}>,
-    private ruleSvc: RuleService,
     private router: Router
     ) {}
 
@@ -28,8 +27,7 @@ export class StartpageComponent implements OnInit {
   }
 
   ruleSetPicked(id: ScenarioName, ruleSetName: RuleSetName) {
-    const rules = this.ruleSvc.getRules(ruleSetName);
-    this.store.dispatch(AppActions.ruleSetPicked({ rules }));
+    this.store.dispatch(AppActions.ruleSetPicked({ rules: ruleSetName }));
     this.store.dispatch(AppActions.scenarioPicked({ scenario: id }));
     this.router.navigate(['/map']);
   }